requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
langchain-openai
langgraph
langgraph_supervisor
